+++ src/data/mutationResults.ts
@@ -4,23 +4,23 @@ import { ApolloExecutionResult } from '../core/types';
 
 // This is part of the public API, people write these functions in `updateQueries`.
 export type MutationQueryReducer<T> = (
-  previousResult: Object,
+  previousResult: Record<string, any>,
   options: {
     mutationResult: ApolloExecutionResult<T>;
     queryName: string | null;
-    queryVariables: Object;
+    queryVariables: Record<string, any>;
   },
-) => Object;
+) => Record<string, any>;
 
 export type MutationQueryReducersMap<T = { [key: string]: any }> = {
   [queryName: string]: MutationQueryReducer<T>;
 };
 
 export type OperationResultReducer = (
-  previousResult: Object,
+  previousResult: Record<string, any>,
   action: ApolloAction,
-  variables: Object,
-) => Object;
+  variables: Record<string, any>,
+) => Record<string, any>;
 
 export type OperationResultReducerMap = {
   [queryId: string]: OperationResultReducer;